+++ datalib/src/main/java/org/vihv/contoureditor/ContourRepository.java
@@ -1,0 +1,6 @@
+//package org.vihv.contoureditor;
+//
+//import org.springframework.data.jpa.repository.JpaRepository;
+//
+//public interface ContourRepository extends JpaRepository<Contour, Long> {
+//}