+++ datalib/src/main/java/org/vihv/contoureditor/EUserNotFound.java
@@ -1,0 +1,4 @@
+package org.vihv.contoureditor;
+
+public class EUserNotFound extends Exception {
+}