+++ datalib/src/main/java/org/vihv/contoureditor/TestApplication.java
@@ -1,0 +1,12 @@
+package org.vihv.contoureditor;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication
+//@EnableJpaRepositories
+@EnableAutoConfiguration
+//@EnableAutoConfiguration(exclude={DataSourceAutoConfiguration.class})
+public class TestApplication {}