+++ datalib/src/main/java/org/vihv/contoureditor/User.java
@@ -1,29 +1,30 @@
 package org.vihv.contoureditor;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
+import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
+import javax.persistence.Id;
+import javax.persistence.Entity;
+import javax.persistence.Table;
 
 @Entity
-@Table(name="user")
+@Table(name="user",
+        uniqueConstraints={
+            @UniqueConstraint(columnNames={"login"})
+        })
 public class User {
-    static final Integer UNDEFINED_ID = null;
+    static final Long UNDEFINED_ID = null;
 
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    public Integer id = null;
+    public Long id = null;
 
     public String login;
     public String password;
-    @ManyToOne
-    public Collection<Contour> contours= new ArrayList<>();
+
+    @OneToMany(cascade = CascadeType.ALL)
+    public List<Contour> contours= new ArrayList<>();
 
 
 }