+++ web/src/main/java/org/vihv/contoureditor/DataConfig.java
@@ -1,0 +1,16 @@
+package org.vihv.contoureditor;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+//@Configuration
+//@ComponentScan({"org.vihv.contoureditor"})
+//@EntityScan(basePackages = "org.vihv.contoureditor")
+//@EnableJpaRepositories(basePackages = "org.vihv.contoureditor")
+////@EnableAutoConfiguration(exclude={DataSourceAutoConfiguration.class})
+//public class DataConfig {
+//}