+++ web/src/test/java/org/vihv/contoureditor/test/AccessTest.java
@@ -1,0 +1,39 @@
+package org.vihv.contoureditor.test;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.web.servlet.MockMvc;
+import org.vihv.contoureditor.Application;
+import org.vihv.contoureditor.IndexController;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@ExtendWith(SpringExtension.class)
+@ContextConfiguration(classes= Application.class)
+@SpringBootTest(classes = { IndexController.class })
+@AutoConfigureMockMvc()
+public class AccessTest {
+
+    @Autowired
+    MockMvc mvc;
+
+    @Test
+    void rootWhenUnauthenticatedThen401() throws Exception {
+        mvc.perform(get("/"))
+                .andExpect(status().isUnauthorized());
+    }
+
+    @Test
+    @WithMockUser
+    public void rootWithMockUserStatusIsOK() throws Exception {
+        mvc.perform(get("/")).andExpect(status().isOk());
+    }
+
+}